+++ apstools/plans/nscan_support.py
@@ -118,7 +118,7 @@ def nscan(detectors, *motor_sets, num=11, per_step=None, md=None):
 # -----------------------------------------------------------------------------
 # :author:    Pete R. Jemian
 # :email:     jemian@anl.gov
-# :copyright: (c) 2017-2022, UChicago Argonne, LLC
+# :copyright: (c) 2017-2023, UChicago Argonne, LLC
 #
 # Distributed under the terms of the Argonne National Laboratory Open Source License.
 #