+++ apstools/utils/log_utils.py
@@ -163,3 +163,13 @@ def stream_log_handler(formatter=None, level="INFO"):
     handler.setLevel(level)
 
     return handler
+
+# -----------------------------------------------------------------------------
+# :author:    Pete R. Jemian
+# :email:     jemian@anl.gov
+# :copyright: (c) 2017-2023, UChicago Argonne, LLC
+#
+# Distributed under the terms of the Argonne National Laboratory Open Source License.
+#
+# The full license is in the file LICENSE.txt, distributed with this software.
+# -----------------------------------------------------------------------------